Rework how shipping consignments are generated.
Instead of the previous approch of auto-generating them on order creation, be kind and let the store decide whe nthey are ready to request a consignment and pack the order. This means there's no mixups during holidays, sick leave, or more.

It also makes it much clearer that the user should resolve their order from within the order it self, and then are immediately presented with a means of getting shipping labels right after.